requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
newsapi-python==0.2.7
flask-wtf==1.2.1
beautifulsoup4==4.12.2
gunicorn==21.2.0 